Cleanup: moved JSON-specific utils to a separate module
