+++ pootle/apps/pootle_language/views.py
@@ -15,9 +15,9 @@ from pootle.core.decorators import get_path_obj, permission_required
 from pootle.core.helpers import (get_export_view_context,
                                  get_overview_context,
                                  get_translation_context)
+from pootle.core.utils.json import jsonify
 from pootle.i18n.gettext import tr_lang
 from pootle_app.views.admin.permissions import admin_permissions
-from pootle_misc.util import jsonify
 
 
 @get_path_obj
