+++ pootle/apps/pootle_misc/util.py
@@ -7,8 +7,6 @@
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
-import json
-
 from functools import wraps
 from importlib import import_module
 
@@ -16,8 +14,6 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.http import HttpResponseBadRequest
 from django.utils import timezone
-from django.utils.encoding import force_unicode
-from django.utils.functional import Promise
 
 # Timezone aware minimum for datetime (if appropriate) (bug 2567)
 from datetime import datetime, timedelta
@@ -25,7 +21,6 @@ datetime_min = datetime.min
 if settings.USE_TZ:
     datetime_min = timezone.make_aware(datetime_min, timezone.utc)
 
-from pootle.core.markup import Markup
 from pootle.core.utils.timezone import make_aware
 
 
@@ -49,31 +44,6 @@ def import_func(path):
 
 def dictsum(x, y):
     return dict((n, x.get(n, 0)+y.get(n, 0)) for n in set(x) | set(y))
-
-
-class PootleJSONEncoder(json.JSONEncoder):
-    """Custom JSON encoder for Pootle.
-
-    This is mostly implemented to avoid calling `force_unicode` all the time on
-    certain types of objects.
-    https://docs.djangoproject.com/en/1.4/topics/serialization/#id2
-    """
-    def default(self, obj):
-        if (isinstance(obj, Promise) or isinstance(obj, Markup) or
-            isinstance(obj, datetime)):
-            return force_unicode(obj)
-
-        return super(PootleJSONEncoder, self).default(obj)
-
-
-def jsonify(obj):
-    """Serialize Python `obj` object into a JSON string."""
-    if settings.DEBUG:
-        indent = 4
-    else:
-        indent = None
-
-    return json.dumps(obj, indent=indent, cls=PootleJSONEncoder)
 
 
 def ajax_required(f):
