+++ pootle/apps/pootle_translationproject/views.py
@@ -20,8 +20,8 @@ from pootle.core.decorators import (get_path_obj, get_resource,
 from pootle.core.helpers import (get_export_view_context,
                                  get_overview_context,
                                  get_translation_context)
+from pootle.core.utils.json import jsonify
 from pootle_app.views.admin.permissions import admin_permissions as admin_perms
-from pootle_misc.util import jsonify
 from staticpages.models import StaticPage
